slight bug fixes
a few changes to logic, mostly on when functions are called. This fixed
the dots being redrawn over and over, live animations are still being
called randomly and need to be addressed. Genres no longer build
repeatedly, but they fail to change on a second hover. Needs to be
revised.